'path_tracking': speed control (constant speed)
Former-commit-id: 41b712e1ac44c30031dd813f3a8dafadb6e52b0e

+++ src/path_tracking/include/Messages.h
@@ -18,6 +18,6 @@ struct PathTrackingMsg
 
 struct Control
 {
-    float speed;
+    int8_t speed;
     float steeringAngle;
 };
